Define `LocalTimelineQuery` and `FederatedTimelineQuery` structs
