+++ Sources/TootSDK/TootClient/TootClient+TimeLine.swift
@@ -39,30 +39,30 @@ public extension TootClient {
     
     /// Retrieves the user's local timeline
     /// - Parameters:
+    ///   - query: a ``LocalTimelineQuery`` struct defining the parameters of the request
     ///   - pageInfo: a PageInfo struct that tells the API how to page the response, typically with a minId set of the highest id you last saw
     ///   - limit: Maximum number of results to return (defaults to 20 on Mastodon with a max of 40)
-    ///   - onlyMedia: Return only statuses with media attachments
     /// - Returns: a PagedResult containing the posts retrieved
    func getLocalTimeline(_ query: LocalTimelineQuery? = nil, _ pageInfo: PagedInfo? = nil, limit: Int? = nil) async throws -> PagedResult<[Post]> {
         let req = HTTPRequestBuilder {
             $0.url = getURL(["api", "v1", "timelines", "public"])
             $0.method = .get
-            $0.query = getQueryParams(pageInfo, limit: limit, onlyMedia: onlyMedia, locality: .local)
+            $0.query = getQueryParams(pageInfo, limit: limit, onlyMedia: query?.onlyMedia, locality: .local)
         }
         return try await getPosts(req, pageInfo, limit)
     }
     
     /// Retrieves the user's federated timeline
     /// - Parameters:
+    ///   - query: a ``FederatedTimelineQuery`` struct defining the parameters of the request
     ///   - pageInfo: a PageInfo struct that tells the API how to page the response, typically with a minId set of the highest id you last saw
     ///   - limit: Maximum number of results to return (defaults to 20 on Mastodon with a max of 40)
-    ///   - onlyMedia: Return only statuses with media attachments
     /// - Returns: a PagedResult containing the posts retrieved
     func getFederatedTimeline(_ query: FederatedTimelineQuery? = nil, _ pageInfo: PagedInfo? = nil, limit: Int? = nil) async throws -> PagedResult<[Post]> {
         let req = HTTPRequestBuilder {
             $0.url = getURL(["api", "v1", "timelines", "public"])
             $0.method = .get
-            $0.query = getQueryParams(pageInfo, limit: limit, onlyMedia: onlyMedia)
+            $0.query = getQueryParams(pageInfo, limit: limit, onlyMedia: query?.onlyMedia)
         }
         return try await getPosts(req, pageInfo, limit)
     }
